progs: Keep only the gftools version of anderson.cpp
